Update README and prepare for release
Updated the description in README.md and changed the project name. For pom.xml, the project version was updated and details including the description, url, and configuration for nexus-staging-maven-plugin were added. Also, a release profile, distribution management settings, license, developers, and scm details were included in preparation for a project release.